+++ packages/api/nest-cli.test.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
-  "entryFile": "main.test",
+  "entryFile": "test/main.test",
   "sourceRoot": "src",
   "compilerOptions": {
     "deleteOutDir": true,
