+++ packages/api/src/infrastructure/notifications/mail/constants/template.type.ts
