+++ packages/api/src/infrastructure/subscription/subscription.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { CityModule } from 'src/application/city/city.module';
-import { NotificationsModule } from 'src/application/notifications/notifications.module';
 import { SubscriptionServiceInterface, SubscriptionServiceLookup } from 'src/application/subscriptions/interfaces/subcription-service.abstract';
 import { GroupSubscriptionRepository, ServiceSubscriptionRepository } from 'src/application/subscriptions/interfaces/subscription-repository.abstract';
+import { SubscriptionService } from 'src/application/subscriptions/subscription.service';
+import { CityModule } from 'src/infrastructure/city/city.module';
 import { SubscriptionRepository } from 'src/infrastructure/subscription/repositories/subscription.repository';
 import { SubscriptionController } from 'src/presentation/subscription/subscription.controller';
 import { SubscriptionDb, SubscriptionSchema } from '../../infrastructure/subscription/schemas/subscription.schema';
-import { SubscriptionService } from './services/subscription.service';
+import { NotificationsModule } from '../notifications/notifications.module';
 
 @Module({
   imports: [
