+++ packages/api/src/test/app.module.test.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { ServeStaticModule } from '@nestjs/serve-static';
-import { CityTestModule } from 'src/application/city/city.module.test';
-import { SubscriptionTestModule } from 'src/application/subscriptions/subscriptions.module.test';
-import { NotificationsTestModule } from './application/notifications/notifications.module.test';
-import { WeatherTestModule } from './application/weather/weather.module.test';
-import { appTestConfig, databaseTestConfig } from './config/test.config';
-import { DatabaseTestModule } from './infrastructure/database/test/database.module.test';
+import { appTestConfig, databaseTestConfig } from 'src/config/test.config';
+import { DatabaseTestModule } from 'src/infrastructure/database/test/database.module.test';
+import { CityTestModule } from './city.module.test';
+import { NotificationsTestModule } from './notifications.module.test';
+import { SubscriptionTestModule } from './subscriptions.module.test';
+import { WeatherTestModule } from './weather.module.test';
 
 @Module({
   imports: [
